renamed references to test.aws_local_stack_access
