+++ test/integration/aws/test_serialization.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from exasol.ds.sandbox.lib.asset_id import AssetId
 from exasol.ds.sandbox.lib.setup_ec2.cf_stack import CloudformationStack
 from exasol.ds.sandbox.lib.setup_ec2.key_file_manager import KeyFileManager
-from test.aws_local_stack_access import AwsLocalStackAccess
+from test.aws.local_stack_access import AwsLocalStackAccess
 import multiprocessing as mp
